Some progress

+++ functions/__init__.h
@@ -1,7 +1,7 @@
 #ifndef __INIT___H
 #define __INIT___H
 
-typedef unsigned int elem_t;
+typedef unsigned long elem_t;
 #define SWAP(a, b) { elem_t tmp = (a); (a) = (b); (b) = tmp; }
 
 #endif
